Struct cpulib::Memory

pub struct Memory {
    pub base_address: usize,
    /* private fields */
}

Represents a memory model with segmented memory blocks. Provides functionality for reading and writing data to specific memory addresses.

Fields
base_address: usize
Implementations

impl Memory

pub fn new(base: usize) -> Self

Creates a new instance of Memory.

Initializes an empty vector of MemorySegment and sets the base address for memory calculations.

Arguments

base - The base address from which all memory addresses will be calculated.

Returns

A new Memory instance with the specified base address.

pub fn read<T: MemoryIO>(&self, address: usize) -> T

Reads a value of type T from memory starting at a given address.

Reads bytes sequentially starting from the address and constructs a value of type T using MemoryIO trait. The number of bytes read is determined by the size of type T.

Type Parameters

T - The type implementing MemoryIO that determines how bytes are read and interpreted.

Arguments

address - The starting address from which to read the bytes.

Returns

A value of type T constructed from the read bytes.

pub fn write<T: MemoryIO>(&mut self, address: usize, value: T)

Writes a value of type T to memory starting at a given address.

Converts the value to a byte array using MemoryIO trait and writes the bytes sequentially to memory. Each byte is written to consecutive memory addresses starting from the specified address.

Type Parameters

T - The type implementing MemoryIO that determines how the value is converted to bytes.

Arguments

address - The starting address at which to write the bytes.

value - The value of type T to write to memory.

pub fn read_vec<T: MemoryIO>( &self, address: usize, number_of_value: usize ) -> Vec<T>

Reads a vector of values of type T from memory starting at a given address.

Reads multiple values sequentially from memory. The number of values read is determined by number_of_value. Each value is read from a block of memory with a size equal to T::size().

Type Parameters

T - The type implementing MemoryIO that determines how bytes are read and interpreted.

Arguments

address - The starting address from which to begin reading values.

number_of_value - The number of values to read.

Returns

A vector of values of type T.

pub fn write_vec<T: MemoryIO + Clone>(&mut self, address: usize, values: Vec<T>)

Writes a vector of values of type T to memory starting at a given address.

Sequentially writes each value in the vector to memory. The address for each subsequent value is offset by T::size() bytes from the previous one.

Type Parameters

T - The type implementing MemoryIO and Clone that determines how values are converted to bytes.

Arguments

address - The starting address at which to begin writing values.

values - The vector of values to write to memory.