Struct cpulib::Registers

pub struct Registers { /* private fields */ }

Represents a collection of registers within a simulated CPU architecture.

This struct includes SIMD registers, general-purpose registers (GPRs), flag registers, and instruction pointers, along with methods to manipulate these registers.

Implementations

impl Registers

pub fn new() -> Self

Creates a new Registers struct with initialized values.

Initializes SIMD registers, GPRs, flag registers, and instruction pointers.

pub fn set_bit( &mut self, reg_type: VecRegName, reg_index: usize, bit_position: usize, value: bool )

Sets a specific bit in a specified SIMD register.

Arguments

reg_type - The type of SIMD register to operate on.

reg_index - The index of the register.

bit_position - The position of the bit to set.

value - The value to set the bit to.

pub fn get_bit( &self, reg_type: VecRegName, reg_index: usize, bit_position: usize ) -> Option<bool>

Gets the value of a specific bit in a specified SIMD register.

Arguments

reg_type - The type of SIMD register to operate on.

reg_index - The index of the register.

bit_position - The position of the bit to retrieve.

Returns

The value of the bit at the specified position, or None if invalid.

pub fn clear(&mut self, reg_index: usize)

Clears all bits in a specified SIMD register.

Arguments

reg_index - The index of the register to clear.

pub fn get_by_sections<T: SectionCompatible>( &self, reg_type: VecRegName, reg_index: usize ) -> Option<Vec<T>>

Retrieves sections of a specified SIMD register as a vector of a specific type.

Type Parameters

T - The type of the sections to be retrieved.

Arguments

reg_type - The type of SIMD register to operate on.

reg_index - The index of the register.

Returns

A vector of T elements representing sections of the register, or None if invalid.

pub fn set_by_sections<T: SectionCompatible>( &mut self, reg_type: VecRegName, reg_index: usize, sections: Vec<T> ) -> bool

Sets sections of a specified SIMD register using a vector of a specific type.

Type Parameters

T - The type of the sections to be set.

Arguments

reg_type - The type of SIMD register to operate on.

reg_index - The index of the register.

sections - The vector of T elements to set in the register.

Returns

true if the operation was successful, false otherwise.

pub fn get_by_selector<T: SectionCompatible>( &self, _reg_type: VecRegName, reg_index: usize, selector: &str ) -> Option<T>

Retrieves a value from a specified SIMD register based on a selector string.

Type Parameters

T - The type of the value to be retrieved.

Arguments

reg_type - The type of SIMD register to operate on.

reg_index - The index of the register.

selector - The string selector determining the range of bits to retrieve.

Returns

The value of the specified type from the register, or None if invalid.

pub fn set_by_selector<T: SectionCompatible>( &mut self, _reg_type: VecRegName, reg_index: usize, selector: &str, value: T ) -> bool

Sets a value in a specified SIMD register based on a selector string.

Type Parameters

T - The type of the value to be set.

Arguments

reg_type - The type of SIMD register to operate on.

reg_index - The index of the register.

selector - The string selector determining the range of bits to set.

value - The value to set in the specified range.

Returns

true if the operation was successful, false otherwise.

pub fn set_gpr_value(&mut self, reg_name: GPRName, value: u64)

Sets the value of a specified general-purpose register.

Handles specific bits based on the register’s type and size.

Arguments

reg_name - The name of the general-purpose register.

value - The value to set the register to.

pub fn get_gpr_value(&self, reg_name: GPRName) -> u64

Retrieves the value of a specified general-purpose register.

Arguments

reg_name - The name of the general-purpose register.

Returns

The current value of the specified register.

pub fn set_flags_value(&mut self, reg_name: FLAGSName, value: u64)

Sets the value of a specified flags register.

Arguments

reg_name - The name of the flags register.

value - The value to set the flags register to.

pub fn get_flags_value(&self, reg_name: FLAGSName) -> u64

Retrieves the value of a specified flags register.

Arguments

reg_name - The name of the flags register.

Returns

The current value of the

pub fn set_ip_value(&mut self, reg_name: IPName, value: u64)

Sets the value of a specified instruction pointer (IP) register.

Handles specific bits based on the IP register’s type and size. This method can be used to set the value of RIP, EIP, or IP registers.

Arguments

reg_name - The name of the IP register, which can be RIP, EIP, or IP.

value - The value to set the IP register to.

pub fn get_ip_value(&self, reg_name: IPName) -> u64

Retrieves the value of a specified instruction pointer (IP) register.

Depending on the type of IP register queried (RIP, EIP, or IP), it returns the respective 64-bit, 32-bit, or 16-bit value.

Arguments

reg_name - The name of the IP register, which can be RIP, EIP, or IP.

Returns

The current value of the specified IP register.